ttt-online-django/requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
